Improve timeout handling in producer/consumer
Previously, if the consumer's overall timeout
passed during it's wait for the next item from
the queue, it would wait until an item shows
up in the queue, so potentially much longer
than the timeout. This change ensures, that
it waits at most the amount of time left before
the overall timeout expires.